data-pipeline/collectors/calendar_collector.py
import requests
from bs4 import BeautifulSoup
from datetime import datetime

class CalendarCollector:
    def __init__(self):
        # 네이버 금융 증시 캘린더 (실적 발표, 배당 등)
        self.cal_url = "https://finance.naver.com/sise/sise_index.naver?code=KOSPI"

    def get_upcoming_events(self, ticker: str):
        """
        해당 종목의 주요 일정(실적 발표, 배당락, 공시 예정 등)을 수집합니다.
        (현재는 네이버 금융 및 DART 일정을 기반으로 한 시뮬레이션 데이터를 포함합니다.)
        """
        # 실제 고도화 시에는 DART API의 'list' 메서드에서 '공시 유형'으로 필터링하여 일정 추출
        events = [
            {"date": "2026-03-25", "event": "정기 주주총회", "impact": "Medium"},
            {"date": "2026-04-15", "event": "1분기 잠정 실적 발표 (예정)", "impact": "High"}
        ]
        return events

# 테스트용
if __name__ == "__main__":
    cal = CalendarCollector()
    print("주요 일정:", cal.get_upcoming_events("005930"))
