data-pipeline/requirements.txt
requests
websockets
python-dotenv
finance-datareader
pandas
OpenDartReader
psycopg2-binary
chromadb
openai
langchain-community
langchain-openai
praw
ntscraper
beautifulsoup4
PyMuPDF
yfinance
pandas_market_calendars
